Moved methods in jiraCommon into the start of a model class
