+++ package.json
@@ -34,6 +34,8 @@
   "jest": {
     "roots": [
       "<rootDir>/test"
-    ]
+    ],
+    "modulePaths":
+    ["<rootDir>"]
   }
 }
